django/requirements.txt
bcrypt==3.1.4
pyotp==2.2.6
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.6.1
Automat==0.7.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
Django==2.0.7
django-extensions==2.0.7
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
pubcontrol==2.4.2
pycparser==2.18
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
txaio==2.10.0
urllib3==1.23
Werkzeug==0.14.1
zope.interface==4.5.0
